rename downloading package to downloadingstate
