+++ app/src/main/java/org/stepic/droid/core/downloadingstate/DownloadingView.kt
@@ -1,4 +1,4 @@
-package org.stepic.droid.core.downloadingprogress
+package org.stepic.droid.core.downloadingstate
 
 interface DownloadingView {
     fun onNewProgressValue(id: Long, portion: Float)
