+++ app/src/main/java/org/stepic/droid/di/section/SectionModule.kt
@@ -2,8 +2,8 @@ package org.stepic.droid.di.section
 
 import dagger.Binds
 import dagger.Module
-import org.stepic.droid.core.downloadingprogress.LessonProgressWatcher
-import org.stepic.droid.core.downloadingprogress.ProgressWatcher
+import org.stepic.droid.core.downloadingstate.LessonProgressWatcher
+import org.stepic.droid.core.downloadingstate.ProgressWatcher
 
 @Module
 interface SectionModule {
