+++ app/src/main/java/org/stepic/droid/ui/fragments/UnitsFragment.java
@@ -27,8 +27,8 @@ import org.stepic.droid.base.App;
 import org.stepic.droid.base.Client;
 import org.stepic.droid.base.FragmentBase;
 import org.stepic.droid.core.LessonSessionManager;
-import org.stepic.droid.core.downloadingprogress.DownloadingPresenter;
-import org.stepic.droid.core.downloadingprogress.DownloadingView;
+import org.stepic.droid.core.downloadingstate.DownloadingPresenter;
+import org.stepic.droid.core.downloadingstate.DownloadingView;
 import org.stepic.droid.core.presenters.DownloadingInteractionPresenter;
 import org.stepic.droid.core.presenters.UnitsLearningProgressPresenter;
 import org.stepic.droid.core.presenters.UnitsPresenter;
